requirements.txt
Pillow==9.2.0
pytest==7.1.3
pytest-cov==4.0.0
